+++ app/src/main/kotlin/com/thesomeshkumar/flixplorer/data/datasource/remote/RemoteDataSourceImpl.kt
@@ -3,16 +3,16 @@ package com.thesomeshkumar.flixplorer.data.datasource.remote
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.thesomeshkumar.flixplorer.data.model.Movie
-import com.thesomeshkumar.flixplorer.data.model.MovieDetailsDTO
-import com.thesomeshkumar.flixplorer.data.model.TVShowDTO
-import com.thesomeshkumar.flixplorer.data.model.TvShowDetailsDTO
-import com.thesomeshkumar.flixplorer.data.paging.AiringTodayTvShowSource
-import com.thesomeshkumar.flixplorer.data.paging.PopularMoviesSource
-import com.thesomeshkumar.flixplorer.data.paging.PopularTvShowSource
-import com.thesomeshkumar.flixplorer.data.paging.TopRatedMoviesSource
-import com.thesomeshkumar.flixplorer.data.paging.TopRatedTvShowSource
-import com.thesomeshkumar.flixplorer.data.paging.UpcomingMoviesSource
+import com.thesomeshkumar.flixplorer.data.model.dto.Movie
+import com.thesomeshkumar.flixplorer.data.model.dto.MovieDetailsDTO
+import com.thesomeshkumar.flixplorer.data.model.dto.TVShowDTO
+import com.thesomeshkumar.flixplorer.data.model.dto.TvShowDetailsDTO
+import com.thesomeshkumar.flixplorer.data.paging.movie.PopularMoviesSource
+import com.thesomeshkumar.flixplorer.data.paging.movie.TopRatedMoviesSource
+import com.thesomeshkumar.flixplorer.data.paging.movie.UpcomingMoviesSource
+import com.thesomeshkumar.flixplorer.data.paging.tvshow.AiringTodayTvShowSource
+import com.thesomeshkumar.flixplorer.data.paging.tvshow.PopularTvShowSource
+import com.thesomeshkumar.flixplorer.data.paging.tvshow.TopRatedTvShowSource
 import com.thesomeshkumar.flixplorer.util.Constants
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
