+++ app/src/main/kotlin/com/thesomeshkumar/flixplorer/data/paging/movie/TopRatedMoviesSource.kt
@@ -1,13 +1,13 @@
-package com.thesomeshkumar.flixplorer.data.paging
+package com.thesomeshkumar.flixplorer.data.paging.movie
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.thesomeshkumar.flixplorer.data.common.RequestErrorHandler
+import com.thesomeshkumar.flixplorer.core.network.error.RequestErrorHandler
 import com.thesomeshkumar.flixplorer.data.datasource.remote.ApiService
-import com.thesomeshkumar.flixplorer.data.model.Movie
+import com.thesomeshkumar.flixplorer.data.model.dto.Movie
 import javax.inject.Inject
 
-class PopularMoviesSource @Inject constructor(
+class TopRatedMoviesSource @Inject constructor(
     private val api: ApiService
 ) : PagingSource<Int, Movie>() {
 
@@ -18,7 +18,7 @@ class PopularMoviesSource @Inject constructor(
     override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Movie> {
         return try {
             val nextPage = params.key ?: 1
-            val popularMovies = api.getPopularMovies(nextPage)
+            val popularMovies = api.getTopRatedMovies(nextPage)
             LoadResult.Page(
                 data = popularMovies.results,
                 prevKey = if (nextPage == 1) null else nextPage - 1,
