+++ app/src/main/kotlin/com/thesomeshkumar/flixplorer/data/paging/tvshow/AiringTodayTvShowSource.kt
@@ -1,13 +1,13 @@
-package com.thesomeshkumar.flixplorer.data.paging
+package com.thesomeshkumar.flixplorer.data.paging.tvshow
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import com.thesomeshkumar.flixplorer.data.common.RequestErrorHandler
+import com.thesomeshkumar.flixplorer.core.network.error.RequestErrorHandler
 import com.thesomeshkumar.flixplorer.data.datasource.remote.ApiService
-import com.thesomeshkumar.flixplorer.data.model.TVShowDTO
+import com.thesomeshkumar.flixplorer.data.model.dto.TVShowDTO
 import javax.inject.Inject
 
-class PopularTvShowSource @Inject constructor(
+class AiringTodayTvShowSource @Inject constructor(
     private val api: ApiService
 ) : PagingSource<Int, TVShowDTO.TVShow>() {
 
@@ -18,9 +18,9 @@ class PopularTvShowSource @Inject constructor(
     override suspend fun load(params: LoadParams<Int>): LoadResult<Int, TVShowDTO.TVShow> {
         return try {
             val nextPage = params.key ?: 1
-            val popularTvShows = api.getPopularTvShows(nextPage)
+            val popularTvShows = api.getAiringTodayTvShows(nextPage)
             LoadResult.Page(
-                data = popularTvShows.results,
+                data = popularTvShows.results.shuffled(),
                 prevKey = if (nextPage == 1) null else nextPage - 1,
                 nextKey = if (popularTvShows.results.isEmpty()) null else popularTvShows.page + 1
             )
